+++ src/components/screens/debug.vue
@@ -7,12 +7,15 @@
   import type { ListItemType, SwipeOptions } from '../elements/list/types';
   import _presets from '../modals/_presets';
   import DebugListComp from './debug-list-comp.vue';
-  import MasterSwitch from '../elements/master-switch.vue';
+  import MasterSwitch from '../elements/switches/master-switch.vue';
   import MdProgress from '../elements/progress/circular-progress.vue';
-  import Switch from '../elements/switch.vue';
+  import Switch from '../elements/switches/switch.vue';
   import Progress from '../elements/progress/linear-progress.vue';
-  import Slider from '../elements/slider.vue';
+  import Slider from '../elements/slider/slider.vue';
   import Chip from '../elements/chip/chip.vue';
+  import Scroller from '../elements/scroller.vue';
+  import MasterSlider from '../elements/slider/master-slider.vue';
+  import { Icon } from '@iconify/vue/dist/iconify.js';
 
   const Lang = window.app.i18n;
   const Option = window.app.options;
@@ -69,15 +72,25 @@
       Use debug options
     </MasterSwitch>
 
-    <MdProgress />
-    <MdProgress :value="progressVal" />
+    <div class="grid-container">
+      <MdProgress />
+      <MdProgress :value="progressVal" />
+
+      <MdProgress :value="progressVal">
+        <MdProgress :diameter="24" :stroke="3" />
+      </MdProgress>
+
+      <MdProgress>
+        <MdProgress :diameter="32" :stroke="3">
+          <MdProgress :diameter="20" :stroke="2">
+            <MdProgress :diameter="8" :stroke="1" />
+          </MdProgress>
+        </MdProgress>
+      </MdProgress>
+    </div>
 
     <Progress />
     <Progress :value="progressVal" />
-
-    <MdProgress :value="progressVal">
-      <MdProgress :diameter="24" :stroke="3" />
-    </MdProgress>
 
     <Slider v-model="progressVal" />
     <Slider :values="[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]" />
@@ -88,6 +101,24 @@
       :change="(v) => Player.setVolume(v)"
     />
     <Slider show-value :step="5" />
+
+    <MasterSlider :change="(v) => Player.setVolume(v)">
+      <Icon icon="material-symbols:volume-up-outline" :width="24" />
+      <span>Volume</span>
+    </MasterSlider>
+
+    <Scroller>
+      Hello there random user, thank you for using this web application, I hope
+      it helped you in some way. What's this? Ah- this is a very long text that
+      is used to test this Scroller component. It's so long, you can't see all
+      of it because of the fact that the your browser window is very small and
+      it can't display it all. It is sad that mobile users won't be able to see
+      all of this meaningless text unless they waited for a long time. Well, if
+      you can see this as a whole, you must have a long-ass monitor or had a
+      multi-monitor setup. Hey you, yeah you, whose using 1920x1080 monitor, you
+      can't un-zoom, that's cheating!. Anyways, thank you very much for using
+      this app 감사합니다!.
+    </Scroller>
 
     <div>
       <Chip
@@ -184,6 +215,12 @@
     align-items: center;
     gap: var(--sm);
   }
+  .grid-container {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(48px, auto));
+    gap: var(--sm);
+  }
+
   .special-wrapper {
     overflow: overlay;
     max-height: 300px;